+++ drone/build.go
@@ -6,14 +6,14 @@ import (
 
 // Build provides build metadata.
 type Build struct {
-	Branch   string    `envconfig:"DRONE_BUILD_BRANCH"`
-	Number   int       `envconfig:"DRONE_BUILD_NUMBER"`
-	Parent   int       `envconfig:"DRONE_BUILD_PARENT"`
-	Event    string    `envconfig:"DRONE_BUILD_EVENT"`
-	Action   string    `envconfig:"DRONE_BUILD_ACTION"`
-	Status   string    `envconfig:"DRONE_BUILD_STATUS"`
-	Created  time.Time `envconfig:"DRONE_BUILD_CREATED"`
-	Started  time.Time `envconfig:"DRONE_BUILD_STARTED"`
-	Finished time.Time `envconfig:"DRONE_BUILD_FINISHED"`
-	Link     string    `envconfig:"DRONE_BUILD_LINK"`
+	Branch   string
+	Number   int
+	Parent   int
+	Event    string
+	Action   string
+	Status   string
+	Created  time.Time
+	Started  time.Time
+	Finished time.Time
+	Link     string
 }