+++ drone/calver.go
@@ -4,11 +4,11 @@ package drone
 // git tag. If the git tag is empty or is not a valid
 // calver, the values will be empty.
 type Calver struct {
-	Version    string `envconfig:"DRONE_CALVER"`
-	Short      string `envconfig:"DRONE_CALVER_SHORT"`
-	MajorMinor string `envconfig:"DRONE_CALVER_MAJOR_MINOR"`
-	Major      string `envconfig:"DRONE_CALVER_MAJOR"`
-	Minor      string `envconfig:"DRONE_CALVER_MINOR"`
-	Micro      string `envconfig:"DRONE_CALVER_MICRO"`
-	Modifier   string `envconfig:"DRONE_CALVER_MODIFIER"`
+	Version    string
+	Short      string
+	MajorMinor string
+	Major      string
+	Minor      string
+	Micro      string
+	Modifier   string
 }