+++ drone/pull_request.go
@@ -2,5 +2,5 @@ package drone
 
 // PullRequest provides the pull request metadata.
 type PullRequest struct {
-	Number int `envconfig:"DRONE_PULL_REQUEST"`
+	Number int
 }