+++ drone/semver.go
@@ -5,12 +5,12 @@ package drone
 // semver, the values will be empty and the error field
 // will be populated with the parsing error.
 type Semver struct {
-	Version    string `envconfig:"DRONE_SEMVER"`
-	Major      string `envconfig:"DRONE_SEMVER_MAJOR"`
-	Minor      string `envconfig:"DRONE_SEMVER_MINOR"`
-	Patch      string `envconfig:"DRONE_SEMVER_PATCH"`
-	PreRelease string `envconfig:"DRONE_SEMVER_PRERELEASE"`
-	Build      string `envconfig:"DRONE_SEMVER_BUILD"`
-	Short      string `envconfig:"DRONE_SEMVER_SHORT"`
-	Error      string `envconfig:"DRONE_SEMVER_ERROR"`
+	Version    string
+	Major      string
+	Minor      string
+	Patch      string
+	PreRelease string
+	Build      string
+	Short      string
+	Error      string
 }